api/pyproject.toml
[tool.poetry]
name = "micro-media"
version = "0.1.0"
description = "A simple micro-service for media management"
authors = ["Alireza Jafari <alirezaja1384@gmail.com>"]
license = ""
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.103.2"
pydantic = "^2.4.2"
sqlalchemy = { extras = ["asyncio"], version = "^2.0.21" }
asyncpg = "^0.28.0"
uvicorn = "^0.23.2"
gunicorn = "^21.2.0"
python-decouple = "^3.8"
fastapi-pagination = "^0.12.10"
fastapi-auth-utils = { git = "https://github.com/Alirezaja1384/fastapi-auth-utils" }
alembic = "^1.12.0"
pyyaml = "^6.0.1"
aioboto3 = "^11.3.0"
python-multipart = "^0.0.6"
pyjwt = {extras = ["crypto"], version = "^2.8.0"}
pillow = "^10.0.1"


[tool.poetry.group.dev.dependencies]
types-aioboto3 = {extras = ["essential"], version = "^11.3.0"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
